requirements.txt
amqp==1.4.9
cassandra-driver==3.1.0
celery==3.1.20
Django==1.8
django-cassandra-engine==0.7.0
django-celery==3.1.17
psycopg2==2.6.1
SQLAlchemy==1.0.12
